Begin Game Room Isolation
Now, starting a game bootstraps it with the necessary state for
play, and loads the gameplay page. The message boxes still need
to be redone with SocketIO rooms.

That will be the logic for passing messages, and saving them for
reloadability. Then there are some edge cases to consider, like
the admin starting without joining the game. Then I can move on
to player customization and landing/lobby design.