Implement Multiple Games
This merges in the feature branch game-creation-and-lobbies.

This creates a barebones landing page, a barebones game lobby,
and refactors the server to handle multiple independent games
instead of just global one. Users can view their current and
completed games or create a new one on the landing page. Creating
one leads to a lobby, where they can join, leave, and invite
friends via the link.