[Server Sockets] Switch from Socket.io's Native Broadcast to a Homebrewn Broadcast
We'd been re-implementing Socket.io's rooms as each Game having an array of Sockets,
because broadcasted events like "notify" needed to be saved with each individual
Player anyways.
May as well go all the way, with all of the broadcasted events in the Player model
that don't need to be saved, so that we can remove the plumbing for the `io` emitter
and have one style for broadcasting.